+++ rs/ic_os/vsock/src/client.rs
@@ -12,7 +12,7 @@ pub use linux::*;
 #[cfg(target_os = "linux")]
 mod linux {
     use super::*;
-    use crate::protocol::{Request, Response, parse_response};
+    use crate::protocol::{Request, Response};
     use std::io::{Read, Write};
     use vsock::{VMADDR_CID_HOST, VsockStream};
 
@@ -66,5 +66,49 @@ mod linux {
 
             parse_response(response_str.as_str())
         }
+    }
+}
+
+/// Parse a response in a json string to a `Response` struct.
+pub fn parse_response(json_str: &str) -> Response {
+    if let Ok(response) = serde_json::from_str::<Response>(json_str) {
+        return response;
+    }
+    Err("Unable to parse host response: ".to_string() + json_str)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::protocol::{HostOSVsockVersion, Payload};
+
+    #[test]
+    fn test_parse_response() {
+        assert_eq!(
+            Ok(Payload::NoPayload),
+            parse_response("{\"Ok\":\"NoPayload\"}")
+        );
+        assert_eq!(
+            Ok(Payload::HostOSVersion("123".to_string())),
+            parse_response("{\"Ok\":{\"HostOSVersion\":\"123\"}}")
+        );
+        assert_eq!(
+            Ok(Payload::HostOSVsockVersion(HostOSVsockVersion {
+                major: 1,
+                minor: 0,
+                patch: 0,
+            })),
+            parse_response(
+                "{\"Ok\":{\"HostOSVsockVersion\":{\"major\":1,\"minor\":0,\"patch\":0}}}"
+            )
+        );
+        assert_eq!(
+            Err("Unable to parse host response: Error response".to_string()),
+            parse_response("Error response")
+        );
+
+        let json_str = r#"{"Ok":"NoPayload"#; // Missing closing brace
+        let response = parse_response(json_str);
+        assert!(response.is_err());
     }
 }
